# openapi.yml
openapi: 3.0.0
info:
  title: TrainWatch API
  description: Predictive maintenance fleet monitoring API
  version: 0.2.0
  contact:
    name: TrainWatch
  license:
    name: MIT

servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.trainwatch.example.com
    description: Production

paths:
  /health:
    get:
      summary: Health check
      tags: [monitoring]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]

  /api/info:
    get:
      summary: API information
      tags: [info]
      responses:
        "200":
          description: API metadata and available endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                  version:
                    type: string
                  endpoints:
                    type: object

  /api/fleet:
    get:
      summary: Complete fleet simulation data
      tags: [fleet]
      description: Returns all train health scores, alerts, and sensor time-series
      responses:
        "200":
          description: Fleet data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FleetData"
        "500":
          description: Server error

  /api/train/{trainId}:
    get:
      summary: Data for a specific train
      tags: [train]
      parameters:
        - name: trainId
          in: path
          required: true
          schema:
            type: string
            example: T-101
      responses:
        "200":
          description: Train data
          content:
            application/json:
              schema:
                type: object
                properties:
                  train:
                    $ref: "#/components/schemas/TrainHealth"
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
                  series:
                    type: object
        "404":
          description: Train not found
        "500":
          description: Server error

  /api/alerts:
    get:
      summary: Query alerts with filters
      tags: [alerts]
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [crit, warn]
        - name: train_id
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Filtered alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
        "500":
          description: Server error

  /api/sensor/{trainId}/{sensor}:
    get:
      summary: Time-series data for one sensor
      tags: [sensor]
      parameters:
        - name: trainId
          in: path
          required: true
          schema:
            type: string
        - name: sensor
          in: path
          required: true
          schema:
            type: string
            enum:
              [
                axle_bearing_temp_c,
                vibration_rms_mm_s,
                brake_pressure_bar,
                traction_motor_current_a,
              ]
      responses:
        "200":
          description: Sensor time-series
          content:
            application/json:
              schema:
                type: object
                properties:
                  sensor:
                    type: string
                  label:
                    type: string
                  unit:
                    type: string
                  series:
                    $ref: "#/components/schemas/TimeSeries"
        "404":
          description: Train or sensor not found
        "500":
          description: Server error

components:
  schemas:
    FleetData:
      type: object
      properties:
        historyEndMs:
          type: integer
          description: Timestamp of last data point (ms since epoch)
        health:
          type: array
          items:
            $ref: "#/components/schemas/TrainHealth"
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        series:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              $ref: "#/components/schemas/TimeSeries"
        totalSamples:
          type: integer
          description: Total telemetry samples in history

    TrainHealth:
      type: object
      properties:
        trainId:
          type: string
          example: T-101
        score:
          type: number
          minimum: 0
          maximum: 100
        status:
          type: string
          enum: [ok, watch, critical]
        activeAlerts:
          type: integer

    Alert:
      type: object
      properties:
        id:
          type: string
        trainId:
          type: string
        sensor:
          type: string
        severity:
          type: string
          enum: [crit, warn]
        detector:
          type: string
          enum: [spike, drift, stuck, limit]
        startedTs:
          type: integer
          description: Timestamp in ms
        endTs:
          type: integer
          description: Timestamp in ms
        active:
          type: boolean
        peak:
          type: number
        recommendation:
          type: string

    TimeSeries:
      type: object
      properties:
        t0:
          type: integer
          description: Start time (unix timestamp)
        dt:
          type: integer
          description: Time delta between points (seconds)
        mean:
          type: array
          items:
            type: number
          description: Hourly mean values
        sev:
          type: array
          items:
            type: integer
            enum: [0, 1, 2]
          description: Max severity per hour (0=ok, 1=warn, 2=crit)
        anom:
          type: array
          items:
            type: integer
          description: Count of anomalous samples per hour